+++ src/apps/reader/Reader.entry.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { withConfig } from "../../core/utils/config";
 import { withUrls } from "../../core/utils/url";
 import { withText } from "../../core/utils/text";
-import Reader, { ReaderType } from "./Reader";
+import Reader, { ReaderType } from "../../components/reader-player/Reader";
 
 const ReaderEntry: React.FC<ReaderType> = ({ identifier, orderId }) => {
   return <Reader identifier={identifier} orderId={orderId} />;
